@@ -11,47 +11,56 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.2.14",
-    "@angular/cdk": "^19.2.19",
-    "@angular/common": "^19.2.14",
-    "@angular/compiler": "^19.2.14",
-    "@angular/core": "^19.2.14",
-    "@angular/forms": "^19.2.14",
-    "@angular/platform-browser": "^19.2.14",
-    "@angular/platform-browser-dynamic": "^19.2.14",
-    "@angular/router": "^19.2.14",
-    "@blackbaud/skyux-lib-addin-client": "^12.1.0",
-    "@skyux/assets": "^12.31.0",
-    "@skyux/config": "^12.31.0",
-    "@skyux/core": "^12.31.0",
-    "@skyux/forms": "^12.31.0",
-    "@skyux/help-inline": "^12.31.0",
-    "@skyux/i18n": "^12.31.0",
-    "@skyux/icon": "^12.31.0",
-    "@skyux/indicators": "^12.31.0",
-    "@skyux/layout": "^12.31.0",
-    "@skyux/modals": "^12.31.0",
-    "@skyux/packages": "^12.31.0",
-    "@skyux/popovers": "^12.31.0",
-    "@skyux/router": "^12.31.0",
-    "@skyux/theme": "^12.31.0",
+    "@angular/animations": "^20.3.16",
+    "@angular/cdk": "^20.2.14",
+    "@angular/common": "^20.3.16",
+    "@angular/compiler": "^20.3.16",
+    "@angular/core": "^20.3.16",
+    "@angular/forms": "^20.3.16",
+    "@angular/platform-browser": "^20.3.16",
+    "@angular/platform-browser-dynamic": "^20.3.16",
+    "@angular/router": "^20.3.16",
+    "@blackbaud/skyux-lib-addin-client": "^13.1.0",
+    "@skyux/assets": "^13.12.0",
+    "@skyux/config": "^13.12.0",
+    "@skyux/core": "^13.12.0",
+    "@skyux/forms": "^13.12.0",
+    "@skyux/help-inline": "^13.12.0",
+    "@skyux/i18n": "^13.12.0",
+    "@skyux/icon": "^13.12.0",
+    "@skyux/indicators": "^13.12.0",
+    "@skyux/layout": "^13.12.0",
+    "@skyux/modals": "^13.12.0",
+    "@skyux/packages": "^13.12.0",
+    "@skyux/pages": "^13.12.0",
+    "@skyux/popovers": "^13.12.0",
+    "@skyux/router": "^13.12.0",
+    "@skyux/theme": "^13.12.0",
     "jwt-decode": "^3.1.2",
     "rxjs": "~7.8.2",
     "tslib": "~2.8.1",
     "zone.js": "~0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.2.15",
-    "@angular/cli": "^19.2.15",
-    "@angular/compiler-cli": "^19.2.14",
-    "@types/jasmine": "~5.1.8",
+    "@angular-devkit/build-angular": "^20.3.15",
+    "@angular-eslint/builder": "^21.2.0",
+    "@angular-eslint/eslint-plugin": "^21.2.0",
+    "@angular-eslint/eslint-plugin-template": "^21.2.0",
+    "@angular-eslint/schematics": "^21.2.0",
+    "@angular-eslint/template-parser": "^21.2.0",
+    "@angular/cli": "^20.3.15",
+    "@angular/compiler-cli": "^20.3.16",
+    "@types/jasmine": "~5.1.15",
+    "@typescript-eslint/eslint-plugin": "^8.54.0",
+    "@typescript-eslint/parser": "^8.54.0",
     "angular-cli-ghpages": "^1.0.7",
-    "jasmine-core": "~5.6.0",
+    "eslint": "^9.39.2",
+    "jasmine-core": "~5.9.0",
     "karma": "~6.4.4",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "2.1.0",
-    "typescript": "~5.7.3"
+    "typescript": "~5.9.3"
   }
 }
